+++ autometa-env.yml
@@ -12,7 +12,7 @@ dependencies:
   - gdown
   - hdbscan
   - hmmer
-  - joblib==1.1.0
+  - joblib
   - numba>=0.47
   - numpy>=1.13
   - pandas>=1.1
