+++ autometa/config/default.config
@@ -62,7 +62,7 @@ bacteria_single_copy_cutoffs = https://${markers:host}/KwanLab/Autometa/main/aut
 archaea_single_copy = https://${markers:host}/KwanLab/Autometa/main/autometa/databases/markers/archaea.single_copy.hmm
 archaea_single_copy_cutoffs = https://${markers:host}/KwanLab/Autometa/main/autometa/databases/markers/archaea.single_copy.cutoffs
 proteins_aa_reps = https://${gtdb:host}/releases/${gtdb:release}/genomic_files_reps/gtdb_proteins_aa_reps.tar.gz
-taxonkit_gtdb_taxdump = https://raw.githubusercontent.com/shenwei356/gtdb-taxdump/releases/latest/download/gtdb-taxdump.tar.gz
+gtdb_taxdmp = https://raw.githubusercontent.com/shenwei356/gtdb-taxdump/releases/latest/download/gtdb-taxdump.tar.gz
 
 
 [checksums]
