+++ autometa/taxonomy/gtdb.py
@@ -9,8 +9,6 @@ import gzip
 import logging
 import os
 import re
-import subprocess
-import shutil
 import string
 import tarfile
 import glob
@@ -24,32 +22,13 @@ from typing import Dict
 import pandas as pd
 import multiprocessing as mp
 
-from autometa.common.utilities import file_length
+from autometa.common.utilities import file_length, is_gz_file
 from autometa.common.external import diamond
 from autometa.taxonomy.database import TaxonomyDatabase
 from autometa.common.exceptions import DatabaseOutOfSyncError
 
 
 logger = logging.getLogger(__name__)
-
-
-def is_gz_file(filepath) -> bool:
-    """
-    Check if the given file is gzipped compressed or not.
-
-    Parameters
-    ----------
-    filepath : str
-        Filepath to check
-
-    Returns
-    -------
-    bool
-        True if file is gzipped else False
-    """
-    # https://stackoverflow.com/a/47080739
-    with open(filepath, "rb") as test_f:
-        return test_f.read(2) == b"\x1f\x8b"
 
 
 def create_gtdb_db(reps_faa: str, dbdir: str) -> str:
@@ -86,14 +65,9 @@ def create_gtdb_db(reps_faa: str, dbdir: str) -> str:
     faa_index: Dict[str, str] = {}
     for genome_protein_faa_filepath in genome_protein_faa_filepaths:
         # Regex to get the genome accession from the file path
-        try:
-            genome_acc_search = re.search(
-                r"GCA_\d+.\d?|GCF_\d+.\d?", genome_protein_faa_filepath
-            )
-        except AttributeError:
-            print(
-                f"Genbank or RefSeq identifier (GCA/GCF) not found in - {genome_protein_faa_filepath}"
-            )
+        genome_acc_search = re.search(
+            r"GCA_\d+.\d?|GCF_\d+.\d?", genome_protein_faa_filepath
+        )
         if not genome_acc_search:
             continue
         faa_index[genome_protein_faa_filepath] = genome_acc_search.group()
@@ -103,13 +77,14 @@ def create_gtdb_db(reps_faa: str, dbdir: str) -> str:
         os.makedirs(dbdir)
 
     logger.debug(f"Merging {len(faa_index):,} faa files.")
-    with open(os.path.join(dbdir, "gtdb_all.faa"), "w") as f_out:
+    combined_faa = os.path.join(dbdir, "gtdb.faa")
+    with open(combined_faa, "w") as f_out:
         for faa_file, acc in faa_index.items():
             with open(faa_file) as f_in:
                 for line in f_in:
                     if line.startswith(">"):
                         seqheader = line.lstrip(">")
-                        line = f">{acc} {seqheader}"
+                        line = f"\n>{acc} {seqheader}"
                     f_out.write(line)
     logger.debug(f"Combined GTDB faa file written to {dbdir}")
     return os.path.join(dbdir, "gtdb.faa")
